+++ resources/views/filament/components/color-picker.blade.php
@@ -41,7 +41,7 @@
                             content: '{{ str($color)->title() }}',
                             theme: $store.theme,
                         }"
-                        class="fi-color hover:ring-gray-500 hover:ring-offset-2 hover:ring-2 transition-all ease-in-out duration-300 cursor-pointer w-6 h-6 bg-primary-500 px-2 py-2 rounded-full">
+                        class="fi-color hover:ring-gray-500 hover:ring-offset-2 hover:ring-2 transition-all ease-in-out duration-300 cursor-pointer w-6 h-6 px-2 py-2 rounded-full">
                     </a>
                 @endforeach
             </div>