+++ CHANGELOG.md
@@ -2,6 +2,37 @@
 
 All notable changes to `Bolt` will be documented in this file
 
+## v4.0.9 - 2025-12-16
+
+**Full Changelog**: https://github.com/lara-zeus/bolt/compare/v4.0.8...v4.0.9
+
+## v4.0.8 - 2025-12-16
+
+### What's Changed
+
+* Remove bg-primary-500 from color picker button by @Abdulmajeed-Jamaan in https://github.com/lara-zeus/bolt/pull/409
+
+**Full Changelog**: https://github.com/lara-zeus/bolt/compare/v4.0.7...v4.0.8
+
+## v4.0.7 - 2025-12-15
+
+### What's Changed
+
+* Fix/support custom toggle colors by @Abdulmajeed-Jamaan in https://github.com/lara-zeus/bolt/pull/408
+
+**Full Changelog**: https://github.com/lara-zeus/bolt/compare/v4.0.6...v4.0.7
+
+## v4.0.6 - 2025-12-07
+
+### What's Changed
+
+* Fix order of field options validation by @Abdulmajeed-Jamaan in https://github.com/lara-zeus/bolt/pull/404
+* Feat: support enums as data source by @Abdulmajeed-Jamaan in https://github.com/lara-zeus/bolt/pull/399
+* improve UX adding the field name and section to the field options modal by @atmonshi in https://github.com/lara-zeus/bolt/pull/403
+* Bump actions/checkout from 5 to 6 by @dependabot[bot] in https://github.com/lara-zeus/bolt/pull/407
+
+**Full Changelog**: https://github.com/lara-zeus/bolt/compare/v4.0.5...v4.0.6
+
 ## v4.0.5 - 2025-10-18
 
 ### What's Changed
@@ -102,6 +133,10 @@ this the change:
     'extensionData' => $extensionData['extInfo']['itemId'] ?? 0,
     'response' => $extensionData['response'],
 ]) !!}
+
+
+
+
 
 
 
@@ -308,6 +343,10 @@ if you need to change this to use another model, add the following in your confi
     //...
     'User' => AnotherUserModel::class,
 ],
+
+
+
+
 
 
 
@@ -916,15 +955,11 @@ and you can remove the fork for `alperenersoy/filament-export`
 **Full Changelog**: https://github.com/lara-zeus/bolt/compare/1.2.3...1.2.4
 ## 1.2.3 - 2023-06-30
 ### What's Changed
-
 - add docs on how to embed the form in any blade file by @atmonshi in https://github.com/lara-zeus/bolt/pull/86
 - more docs by @atmonshi in https://github.com/lara-zeus/bolt/pull/87
 - update core by @atmonshi in https://github.com/lara-zeus/bolt/pull/88
-
 **Full Changelog**: https://github.com/lara-zeus/bolt/compare/1.2.2...1.2.3
-
 ## 1.2.2 - 2023-06-29
-
 ### What's Changed
 
 - fix: pass ext data instead of depending on request by @atmonshi in https://github.com/lara-zeus/bolt/pull/85